+++ doc/database/readme.md
@@ -17,24 +17,6 @@ To test your migration without clearing your database, run `sqlx migrate run`.
 
 Other developers will get your migrations when they run `dev-migrate-schema`.
 
-### Keeping local backups of the database
-
-**For a happier and healthier life**, consider backing up your database before running experimental migrations. "Undoing" a migration is difficult, so backing up and restoring can be the way to go. Overall, this is much quicker than running `dev-migrate-schema` and `dev-migrate-data` everytime you want to update and rerun you migration. The below should all be run _with your database running_.
-
-To create a backup of the database run:
-
-```zsh
-pg_dump dailp > backup.sql
-```
-
-To restore this backup run:
-
-```zsh
-dropdb dailp
-createdb -T template0 dailp
-psql dailp < backup.sql
-```
-
 ## Abbreviations in this Folder
 
 Most of our columns are `not null`, which is long to write so we introduced shorthand for describing database columns.
