+++ graphql/schema.graphql
@@ -195,10 +195,6 @@ input AnnotatedFormUpdate {
 	Possible update to commentary
 	"""
 	commentary: String
-	"""
-	Updated segments
-	"""
-	segments: [MorphemeSegmentUpdate!]
 }
 
 """
@@ -403,31 +399,9 @@ type Date {
 	"""
 	year: Int!
 	"""
-	The month of this date
-	"""
-	month: Int!
-	"""
-	The day of this date
-	"""
-	day: Int!
-	"""
 	Formatted version of the date for humans to read
 	"""
 	formattedDate: String!
-}
-
-input DateInput {
-	day: Int!
-	month: Int!
-	year: Int!
-}
-
-"""
-Delete a contributor attribution for a document based on the two ids
-"""
-input DeleteContributorAttribution {
-	documentId: UUID!
-	contributorId: UUID!
 }
 
 type DocumentCollection {
@@ -446,16 +420,6 @@ type DocumentCollection {
 	pass that to the dataloader below.
 	"""
 	documents: [DocumentReference!]!
-}
-
-"""
-Used for updating document metadata.
-All fields except id are optional.
-"""
-input DocumentMetadataUpdate {
-	id: UUID!
-	title: String
-	writtenAt: DateInput
 }
 
 type DocumentPage {
@@ -646,29 +610,6 @@ type MorphemeReference {
 }
 
 """
-A single unit of meaning and its gloss which can be edited.
-"""
-input MorphemeSegmentUpdate {
-	"""
-	Which Cherokee representation system is this segment written with?
-	"""
-	system: CherokeeOrthography
-	"""
-	Source language representation of this segment.
-	"""
-	morpheme: String!
-	"""
-	Target language representation of this segment.
-	"""
-	gloss: String!
-	"""
-	This field determines what character should separate this segment from
-	the next one when reconstituting the full segmentation string.
-	"""
-	role: WordSegmentRole!
-}
-
-"""
 A concrete representation of a particular functional morpheme.
 """
 type MorphemeTag {
@@ -718,14 +659,6 @@ type Mutation {
 	"""
 	apiVersion: String!
 	"""
-	Mutation for adding/changing contributor attributions
-	"""
-	updateContributorAttribution(contribution: UpdateContributorAttribution!): UUID!
-	"""
-	Mutation for deleting contributor attributions
-	"""
-	deleteContributorAttribution(contribution: DeleteContributorAttribution!): UUID!
-	"""
 	Mutation for paragraph and translation editing
 	"""
 	updateParagraph(paragraph: ParagraphUpdate!): UUID!
@@ -740,8 +673,11 @@ type Mutation {
 	Decide if a piece audio should be included in edited collection
 	"""
 	curateWordAudio(input: CurateWordAudioInput!): AnnotatedForm!
+	"""
+	Attach audio that has already been uploaded to S3 to a particular word
+	Assumes user requesting mutation recoreded the audio
+	"""
 	attachAudioToWord(input: AttachAudioToWordInput!): AnnotatedForm!
-	updateDocumentMetadata(document: DocumentMetadataUpdate!): UUID!
 }
 
 """
@@ -922,15 +858,6 @@ entities without requiring a central allocating authority.
 * [RFC4122: A Universally Unique IDentifier (UUID) URN Namespace](http://tools.ietf.org/html/rfc4122)
 """
 scalar UUID
-
-"""
-Update the contributor attribution for a document
-"""
-input UpdateContributorAttribution {
-	documentId: UUID!
-	contributorId: UUID!
-	contributionRole: String!
-}
 
 """
 A user record, for a contributor, editor, etc.
