+++ graphql/src/query.rs
@@ -2,7 +2,7 @@
 
 use dailp::{
     slugify_ltree, user::AddBookmark, AnnotatedForm, AttachAudioToWordInput, CollectionChapter,
-    CurateWordAudioInput, DeleteContributorAttribution, DocumentMetadataUpdate, UpdateContributorAttribution, Uuid,
+    CurateWordAudioInput, Uuid,
 };
 use itertools::Itertools;
 
@@ -328,38 +328,6 @@ impl Mutation {
         "1.0"
     }
 
-    /// Mutation for adding/changing contributor attributions
-    #[graphql(
-        guard = "GroupGuard::new(UserGroup::Editors).or(GroupGuard::new(UserGroup::Contributors))"
-    )]
-    async fn update_contributor_attribution(
-        &self,
-        context: &Context<'_>,
-        contribution: UpdateContributorAttribution,
-    ) -> FieldResult<Uuid> {
-        Ok(context
-            .data::<DataLoader<Database>>()?
-            .loader()
-            .update_contributor_attribution(contribution)
-            .await?)
-    }
-
-    ///Mutation for deleting contributor attributions
-    #[graphql(
-        guard = "GroupGuard::new(UserGroup::Editors).or(GroupGuard::new(UserGroup::Contributors))"
-    )]
-    async fn delete_contributor_attribution(
-        &self,
-        context: &Context<'_>,
-        contribution: DeleteContributorAttribution,
-    ) -> FieldResult<Uuid> {
-        Ok(context
-            .data::<DataLoader<Database>>()?
-            .loader()
-            .delete_contributor_attribution(contribution)
-            .await?)
-    }
-
     /// Mutation for paragraph and translation editing
     #[graphql(guard = "GroupGuard::new(UserGroup::Editors)")]
     async fn update_paragraph(
@@ -481,19 +449,6 @@ impl Mutation {
             .data::<DataLoader<Database>>()?
             .loader()
             .word_by_id(&word_id)
-            .await?)
-    }
-
-    #[graphql(guard = "GroupGuard::new(UserGroup::Editors)")]
-    async fn update_document_metadata(
-        &self,
-        context: &Context<'_>,
-        document: DocumentMetadataUpdate,
-    ) -> FieldResult<Uuid> {
-        Ok(context
-            .data::<DataLoader<Database>>()?
-            .loader()
-            .update_document_metadata(document)
             .await?)
     }
 }
