+++ netlify.toml
@@ -2,6 +2,3 @@
 base = "website/"
 publish = "public/"
 command = "yarn build"
-
-[build.environment]
-NODE_VERSION = "18"
