+++ terraform/website.nix
@@ -30,6 +30,7 @@ with builtins; {
       in {
         lifecycle.prevent_destroy = true;
         name = "dailp";
+        tags = config.setup.global_tags;
         description = "Digital Archive of Indigenous Language Persistence";
         repository = lib.toLower (getEnv "GIT_REPOSITORY_URL");
         oauth_token = getEnv "OAUTH_TOKEN";
@@ -60,7 +61,6 @@ with builtins; {
             DAILP_AWS_REGION = config.provider.aws.region;
             DAILP_USER_POOL = "\${aws_cognito_user_pool.main.id}";
             DAILP_USER_POOL_CLIENT = "\${aws_cognito_user_pool_client.main.id}";
-            DAILP_IDENTITY_POOL = getEnv "DAILP_IDENTITY_POOL";
             TF_STAGE = config.setup.stage;
             VITE_DEPLOYMENT_ENV = config.setup.stage;
           };
