--- types/queries/delete_contributor_attribution.sql
@@ -1,2 +1,0 @@
-DELETE FROM contributor_attribution
-WHERE document_id = $1 AND contributor_id = $2;
