+++ types/queries/update_word.sql
@@ -10,4 +10,3 @@ update word set
             else commentary
         end
 where id = $1
-returning word.document_id;
