+++ types/src/date.rs
@@ -29,47 +29,8 @@ impl Date {
         self.0.year()
     }
 
-    /// The month of this date
-    pub async fn month(&self) -> u32 {
-        use chrono::Datelike as _;
-        self.0.month()
-    }
-
-    /// The day of this date
-    pub async fn day(&self) -> u32 {
-        use chrono::Datelike as _;
-        self.0.day()
-    }
-
     /// Formatted version of the date for humans to read
     pub async fn formatted_date(&self) -> String {
         self.0.format("%B %e, %Y").to_string()
-    }
-}
-
-#[derive(async_graphql::InputObject)]
-pub struct DateInput {
-    day: u32,
-    month: u32,
-    year: i32,
-}
-
-impl Into<Date> for &DateInput {
-    fn into(self) -> Date {
-        Date::from_ymd(self.year, self.month, self.day)
-    }
-}
-
-use chrono::Datelike;
-impl DateInput {
-    // Creates a new DateInput from a day, month, and year.
-    pub fn new(day: u32, month: u32, year: i32) -> Self {
-        Self { day, month, year }
-    }
-
-    //Creates a new DateInput from a Date object.
-    pub fn parse_date_object(d: Date) -> Self {
-        let nd = d.0;
-        Self::new(nd.day(), nd.month(), nd.year())
     }
 }
