+++ types/src/morpheme.rs
@@ -163,17 +163,3 @@ impl WordSegment {
         }
     }
 }
-
-/// A single unit of meaning and its gloss which can be edited.
-#[derive(async_graphql::InputObject)]
-pub struct MorphemeSegmentUpdate {
-    /// Which Cherokee representation system is this segment written with?
-    pub system: Option<CherokeeOrthography>,
-    /// Source language representation of this segment.
-    pub morpheme: String,
-    /// Target language representation of this segment.
-    pub gloss: String,
-    /// This field determines what character should separate this segment from
-    /// the next one when reconstituting the full segmentation string.
-    pub role: WordSegmentRole,
-}
