+++ website/package.json
@@ -2,7 +2,6 @@
   "name": "dailp-website",
   "version": "0.1.0",
   "description": "Front-end for exploring Cherokee documents from DAILP",
-  "type": "module",
   "scripts": {
     "dev:server": "vite --port 3001",
     "dev:generate": "graphql-codegen --watch",
@@ -41,12 +40,10 @@
     "pure-react-carousel": "^1.30.1",
     "query-string": "^7.1.1",
     "react": "17",
-    "react-date-picker": "^10.5.0",
     "react-device-detect": "^2.2.2",
     "react-dom": "17",
     "react-helmet": "^6.0.0",
     "react-icons": "^4.7",
-    "react-select": "^5.4.0",
     "react-zoom-pan-pinch": "^2.1.3",
     "reakit": "^1.3.11",
     "urql": "^3.0",
@@ -85,7 +82,8 @@
     "typescript": "^4.6.3",
     "typography": "^0.16.19",
     "vite": "^3",
-    "vite-plugin-checker": "^0.5"
+    "vite-plugin-checker": "^0.5",
+    "vite-tsconfig-paths": "^4.0"
   },
   "browserslist": [
     "supports es6-module",
