+++ website/src/edit-word-feature.tsx
@@ -5,20 +5,17 @@ import {
   unstable_Form as Form,
   unstable_FormInput as FormInput,
   unstable_FormLabel as FormLabel,
-  unstable_FormSubmitButton as FormSubmitButton,
 } from "reakit"
-import * as Dailp from "src/graphql/dailp"
+import { unstable_FormSubmitButton as FormSubmitButton } from "reakit"
 import { IconButton } from "./components"
 import { IconTextButton } from "./components/button"
 import * as css from "./edit-word-feature.css"
-import { useForm } from "./edit-word-form-context"
-import { usePreferences } from "./preferences-context"
+import { useForm } from "./form-context"
+import * as Dailp from "./graphql/dailp"
 
 /** Button that allows user to enter edit mode in the word panel, and edit fields of a word. */
 export const EditButton = () => {
   const { form, isEditing, setIsEditing } = useForm()
-
-  const { cherokeeRepresentation } = usePreferences()
 
   return (
     <Form {...form} className={css.form}>
@@ -49,13 +46,7 @@ export const EditButton = () => {
           icon={<HiPencilAlt />}
           className={css.editPanelButton}
           onClick={() => {
-            if (cherokeeRepresentation === Dailp.CherokeeOrthography.Taoc) {
-              setIsEditing(true)
-            } else {
-              alert(
-                "Currently, only the linguistic analysis using terms from Tone and Accent in Oklahoma Cherokee (TAOC) is supported for editing. Please update your Cherokee description style in the display settings."
-              )
-            }
+            setIsEditing(true)
           }}
         >
           Edit
