+++ website/src/graphql/dailp/index.ts
@@ -191,8 +191,6 @@ export type AnnotatedFormUpdate = {
   readonly commentary: InputMaybe<Scalars["String"]>
   /** Unique identifier of the form */
   readonly id: Scalars["UUID"]
-  /** Updated segments */
-  readonly segments: InputMaybe<ReadonlyArray<MorphemeSegmentUpdate>>
   /** Possible update to source content */
   readonly source: InputMaybe<Scalars["String"]>
 }
@@ -340,26 +338,10 @@ export type CurateWordAudioInput = {
 
 export type Date = {
   readonly __typename?: "Date"
-  /** The day of this date */
-  readonly day: Scalars["Int"]
   /** Formatted version of the date for humans to read */
   readonly formattedDate: Scalars["String"]
-  /** The month of this date */
-  readonly month: Scalars["Int"]
   /** The year of this date */
   readonly year: Scalars["Int"]
-}
-
-export type DateInput = {
-  readonly day: Scalars["Int"]
-  readonly month: Scalars["Int"]
-  readonly year: Scalars["Int"]
-}
-
-/** Delete a contributor attribution for a document based on the two ids */
-export type DeleteContributorAttribution = {
-  readonly contributorId: Scalars["UUID"]
-  readonly documentId: Scalars["UUID"]
 }
 
 export type DocumentCollection = {
@@ -375,16 +357,6 @@ export type DocumentCollection = {
   readonly name: Scalars["String"]
   /** URL-ready slug for this collection, generated from the name */
   readonly slug: Scalars["String"]
-}
-
-/**
- * Used for updating document metadata.
- * All fields except id are optional.
- */
-export type DocumentMetadataUpdate = {
-  readonly id: Scalars["UUID"]
-  readonly title: InputMaybe<Scalars["String"]>
-  readonly writtenAt: InputMaybe<DateInput>
 }
 
 export type DocumentPage = {
@@ -530,21 +502,6 @@ export type MorphemeReference = {
   readonly morpheme: Scalars["String"]
 }
 
-/** A single unit of meaning and its gloss which can be edited. */
-export type MorphemeSegmentUpdate = {
-  /** Target language representation of this segment. */
-  readonly gloss: Scalars["String"]
-  /** Source language representation of this segment. */
-  readonly morpheme: Scalars["String"]
-  /**
-   * This field determines what character should separate this segment from
-   * the next one when reconstituting the full segmentation string.
-   */
-  readonly role: WordSegmentRole
-  /** Which Cherokee representation system is this segment written with? */
-  readonly system: InputMaybe<CherokeeOrthography>
-}
-
 /** A concrete representation of a particular functional morpheme. */
 export type MorphemeTag = {
   readonly __typename?: "MorphemeTag"
@@ -593,12 +550,7 @@ export type Mutation = {
   readonly attachAudioToWord: AnnotatedForm
   /** Decide if a piece audio should be included in edited collection */
   readonly curateWordAudio: AnnotatedForm
-  /** Mutation for deleting contributor attributions */
-  readonly deleteContributorAttribution: Scalars["UUID"]
   readonly updateAnnotation: Scalars["Boolean"]
-  /** Mutation for adding/changing contributor attributions */
-  readonly updateContributorAttribution: Scalars["UUID"]
-  readonly updateDocumentMetadata: Scalars["UUID"]
   readonly updatePage: Scalars["Boolean"]
   /** Mutation for paragraph and translation editing */
   readonly updateParagraph: Scalars["UUID"]
@@ -617,20 +569,8 @@ export type MutationCurateWordAudioArgs = {
   input: CurateWordAudioInput
 }
 
-export type MutationDeleteContributorAttributionArgs = {
-  contribution: DeleteContributorAttribution
-}
-
 export type MutationUpdateAnnotationArgs = {
   data: Scalars["JSON"]
-}
-
-export type MutationUpdateContributorAttributionArgs = {
-  contribution: UpdateContributorAttribution
-}
-
-export type MutationUpdateDocumentMetadataArgs = {
-  document: DocumentMetadataUpdate
 }
 
 export type MutationUpdatePageArgs = {
@@ -829,13 +769,6 @@ export type SourceAttribution = {
   readonly link: Scalars["String"]
   /** Name of the source, i.e. "The Newberry Library" */
   readonly name: Scalars["String"]
-}
-
-/** Update the contributor attribution for a document */
-export type UpdateContributorAttribution = {
-  readonly contributionRole: Scalars["String"]
-  readonly contributorId: Scalars["UUID"]
-  readonly documentId: Scalars["UUID"]
 }
 
 /** A user record, for a contributor, editor, etc. */
@@ -1059,9 +992,6 @@ export type DocumentContentsQuery = { readonly __typename?: "Query" } & {
                                   >
                                 }
                             >
-                            readonly position: {
-                              readonly __typename?: "PositionInDocument"
-                            } & Pick<PositionInDocument, "documentId">
                           })
                       | { readonly __typename: "LineBreak" }
                     >
@@ -1134,9 +1064,6 @@ export type DocumentContentsQuery = { readonly __typename?: "Query" } & {
                   >
                 }
             >
-            readonly position: {
-              readonly __typename?: "PositionInDocument"
-            } & Pick<PositionInDocument, "documentId">
           }
       >
     }
@@ -1154,14 +1081,6 @@ export type AudioSliceFieldsFragment = {
   | "endTime"
   | "includeInEditedCollection"
 >
-
-export type DocFormFieldsFragment = {
-  readonly __typename?: "AnnotatedDoc"
-} & Pick<AnnotatedDoc, "id" | "title"> & {
-    readonly date: Maybe<
-      { readonly __typename?: "Date" } & Pick<Date, "day" | "month" | "year">
-    >
-  }
 
 export type FormFieldsFragment = {
   readonly __typename?: "AnnotatedForm"
@@ -1224,10 +1143,6 @@ export type FormFieldsFragment = {
             { readonly __typename?: "User" } & Pick<User, "id" | "displayName">
           >
         }
-    >
-    readonly position: { readonly __typename?: "PositionInDocument" } & Pick<
-      PositionInDocument,
-      "documentId"
     >
   }
 
@@ -1548,9 +1463,6 @@ export type DocSliceQuery = { readonly __typename?: "Query" } & {
                     >
                   }
               >
-              readonly position: {
-                readonly __typename?: "PositionInDocument"
-              } & Pick<PositionInDocument, "documentId">
             }
         >
       }
@@ -1689,10 +1601,6 @@ export type UpdateWordMutation = { readonly __typename?: "Mutation" } & {
             >
           }
       >
-      readonly position: { readonly __typename?: "PositionInDocument" } & Pick<
-        PositionInDocument,
-        "documentId"
-      >
     }
 }
 
@@ -1781,74 +1689,6 @@ export type GetDocShortNameQuery = { readonly __typename?: "Query" } & Pick<
   "getDocShortName"
 >
 
-export type AddBookmarkMutationVariables = Exact<{
-  bookmark: AddBookmark
-}>
-
-export type AddBookmarkMutation = { readonly __typename?: "Mutation" } & Pick<
-  Mutation,
-  "addBookmark"
->
-
-export type GetBookmarksQueryVariables = Exact<{ [key: string]: never }>
-
-export type GetBookmarksQuery = { readonly __typename?: "Query" } & Pick<
-  Query,
-  "getBookmarks"
->
-
-export type GetDocShortNameQueryVariables = Exact<{
-  docId: Scalars["UUID"]
-}>
-
-export type GetDocShortNameQuery = { readonly __typename?: "Query" } & Pick<
-  Query,
-  "getDocShortName"
->
-
-export type UpdateParagraphMutationVariables = Exact<{
-  paragraph: ParagraphUpdate
-}>
-
-export type UpdateParagraphMutation = {
-  readonly __typename?: "Mutation"
-} & Pick<Mutation, "updateParagraph">
-
-export type UpdateContributorAttributionMutationVariables = Exact<{
-  contribution: UpdateContributorAttribution
-}>
-
-export type UpdateContributorAttributionMutation = {
-  readonly __typename?: "Mutation"
-} & Pick<Mutation, "updateContributorAttribution">
-
-export type DeleteContributorAttributionMutationVariables = Exact<{
-  contribution: DeleteContributorAttribution
-}>
-
-export type DeleteContributorAttributionMutation = {
-  readonly __typename?: "Mutation"
-} & Pick<Mutation, "deleteContributorAttribution">
-
-export type UpdateDocumentMetadataMutationVariables = Exact<{
-  document: DocumentMetadataUpdate
-}>
-
-export type UpdateDocumentMetadataMutation = {
-  readonly __typename?: "Mutation"
-} & Pick<Mutation, "updateDocumentMetadata">
-
-export const DocFormFieldsFragmentDoc = gql`
-  fragment DocFormFields on AnnotatedDoc {
-    id
-    title
-    date {
-      day
-      month
-      year
-    }
-  }
-`
 export const AudioSliceFieldsFragmentDoc = gql`
   fragment AudioSliceFields on AudioSlice {
     sliceId
@@ -1890,9 +1730,6 @@ export const FormFieldsFragmentDoc = gql`
         id
         displayName
       }
-    }
-    position {
-      documentId
     }
   }
 `
@@ -2463,57 +2300,4 @@ export function useGetDocShortNameQuery(
     query: GetDocShortNameDocument,
     ...options,
   })
-}
-
-export const UpdateParagraphDocument = gql`
-  mutation UpdateParagraph($paragraph: ParagraphUpdate!) {
-    updateParagraph(paragraph: $paragraph)
-  }
-`
-
-export function useUpdateParagraphMutation() {
-  return Urql.useMutation<
-    UpdateParagraphMutation,
-    UpdateParagraphMutationVariables
-  >(UpdateParagraphDocument)
-}
-export const UpdateContributorAttributionDocument = gql`
-  mutation UpdateContributorAttribution(
-    $contribution: UpdateContributorAttribution!
-  ) {
-    updateContributorAttribution(contribution: $contribution)
-  }
-`
-
-export function useUpdateContributorAttributionMutation() {
-  return Urql.useMutation<
-    UpdateContributorAttributionMutation,
-    UpdateContributorAttributionMutationVariables
-  >(UpdateContributorAttributionDocument)
-}
-export const DeleteContributorAttributionDocument = gql`
-  mutation DeleteContributorAttribution(
-    $contribution: DeleteContributorAttribution!
-  ) {
-    deleteContributorAttribution(contribution: $contribution)
-  }
-`
-
-export function useDeleteContributorAttributionMutation() {
-  return Urql.useMutation<
-    DeleteContributorAttributionMutation,
-    DeleteContributorAttributionMutationVariables
-  >(DeleteContributorAttributionDocument)
-}
-export const UpdateDocumentMetadataDocument = gql`
-  mutation UpdateDocumentMetadata($document: DocumentMetadataUpdate!) {
-    updateDocumentMetadata(document: $document)
-  }
-`
-
-export function useUpdateDocumentMetadataMutation() {
-  return Urql.useMutation<
-    UpdateDocumentMetadataMutation,
-    UpdateDocumentMetadataMutationVariables
-  >(UpdateDocumentMetadataDocument)
 }
