+++ website/src/graphql/dailp/queries.graphql
@@ -80,16 +80,6 @@ fragment AudioSliceFields on AudioSlice {
   includeInEditedCollection
 }
 
-fragment DocFormFields on AnnotatedDoc {
-  id
-  title
-  date {
-    day
-    month
-    year
-  }
-}
-
 fragment FormFields on AnnotatedForm {
   id
   index
@@ -120,9 +110,6 @@ fragment FormFields on AnnotatedForm {
       id
       displayName
     }
-  }
-  position {
-    documentId
   }
 }
 
@@ -389,24 +376,4 @@ query GetBookmarks {
 
 query GetDocShortName($docId: UUID!) {
   getDocShortName(docId: $docId)
-}
-
-mutation UpdateParagraph($paragraph: ParagraphUpdate!) {
-  updateParagraph(paragraph: $paragraph)
-}
-
-mutation UpdateContributorAttribution(
-  $contribution: UpdateContributorAttribution!
-) {
-  updateContributorAttribution(contribution: $contribution)
-}
-
-mutation DeleteContributorAttribution(
-  $contribution: DeleteContributorAttribution!
-) {
-  deleteContributorAttribution(contribution: $contribution)
-}
-
-mutation UpdateDocumentMetadata($document: DocumentMetadataUpdate!) {
-  updateDocumentMetadata(document: $document)
 }
