+++ website/src/pages/documents/document.css.ts
@@ -235,48 +235,6 @@ export const mobileWordPanel = style([
   },
 ])
 
-// export const hideScrollTop = style({
-//   display: "flex",
-//   flexFlow: "column nowrap",
-//   justifyContent: "flex-end",
-//   alignContent: "center",
-//   alignItems: "center",
-//   alignSelf: "flex-end",
-//   position: "sticky",
-//   zIndex: 1,
-//   top: `calc(100vh + 100px)`,
-//   right: 0,
-//   padding: vspace.quarter,
-//   color: colors.secondaryContrast,
-//   outlineColor: colors.secondaryContrast,
-//   backgroundColor: colors.secondary,
-//   visibility: "visible",
-//   opacity: 1,
-//   transition: "top 0.8s",
-// })
-
-// export const showScrollTop = style(
-//   {
-//   display: "flex",
-//   flexFlow: "column nowrap",
-//   justifyContent: "flex-end",
-//   alignContent: "center",
-//   alignItems: "center",
-//   alignSelf: "flex-end",
-//   position: "sticky",
-//   zIndex: 1,
-//   top: `calc(100vh - 100px)`,
-//   right: 0,
-//   padding: vspace.quarter,
-//   color: colors.secondaryContrast,
-//   outlineColor: colors.secondaryContrast,
-//   backgroundColor: colors.secondary,
-//   visibility: "visible",
-//   opacity: 1,
-//   transition: "top 0.5s",
-//   }
-// )
-
 export const scrollTop = style({
   display: "flex",
   flexFlow: "column nowrap",
@@ -286,6 +244,7 @@ export const scrollTop = style({
   alignSelf: "flex-end",
   position: "sticky",
   zIndex: 1,
+  top: `calc(100vh - 100px)`,
   right: 0,
   padding: vspace.quarter,
   color: colors.secondaryContrast,
@@ -303,28 +262,3 @@ export const BookmarkButton = style([
     marginLeft: hspace.small,
   },
 ])
-
-export const hideScrollTop = style([
-  scrollTop,
-  {
-    top: `calc(100vh + 100px)`,
-    visibility: "visible",
-    opacity: 1,
-    transition: "top 0.8s",
-  },
-])
-
-export const showScrollTop = style([
-  scrollTop,
-  {
-    top: `calc(100vh - 100px)`,
-    visibility: "visible",
-    opacity: 1,
-    transition: "top 0.5s",
-  },
-])
-
-export const noScrollTop = style({
-  visibility: "hidden",
-  opacity: 0,
-})
