+++ website/src/panel-layout.tsx
@@ -1,27 +1,17 @@
-import { groupBy } from "lodash"
 import React, { ReactNode } from "react"
 import { AiFillSound } from "react-icons/ai/index"
 import { GrDown, GrUp } from "react-icons/gr/index"
 import { IoEllipsisHorizontalCircle } from "react-icons/io5/index"
 import { MdClose, MdNotes, MdRecordVoiceOver } from "react-icons/md/index"
-import { OnChangeValue } from "react-select"
-import {
-  Disclosure,
-  DisclosureContent,
-  useDisclosureState,
-} from "reakit"
-import {
-  unstable_Form as Form,
-  unstable_FormInput as FormInput,
-} from "reakit"
-import * as Dailp from "src/graphql/dailp"
+import { Disclosure, DisclosureContent, useDisclosureState } from "reakit"
+import { unstable_Form as Form } from "reakit"
 import { useCredentials } from "./auth"
 import { AudioPlayer, IconButton } from "./components"
-import { CustomCreatable } from "./components/creatable"
 import { EditWordAudio } from "./components/edit-word-audio"
 import { EditButton, EditWordFeature } from "./edit-word-feature"
-import { useForm } from "./edit-word-form-context"
 import { content } from "./footer.css"
+import { useForm } from "./form-context"
+import * as Dailp from "./graphql/dailp"
 import * as css from "./panel-layout.css"
 import ParagraphPanel from "./paragraph-panel"
 import { usePreferences } from "./preferences-context"
@@ -29,8 +19,8 @@ import { TranslatedParagraph } from "./segment"
 import { VerticalMorphemicSegmentation } from "./word-panel"
 
 enum PanelType {
-  WordPanel,
   EditWordPanel,
+  WordPanel,
 }
 
 export type PanelSegment = Dailp.FormFieldsFragment | TranslatedParagraph
@@ -40,7 +30,7 @@ export interface PanelDetails {
   setCurrContents: (currContents: PanelSegment | null) => void
 }
 
-// Displays the right-side panel information of the currently selected segment.
+/** Displays the right-side panel information of the currently selected word. */
 export const PanelLayout = (p: {
   segment: PanelSegment | null
   setContent: (content: PanelSegment | null) => void
@@ -52,39 +42,8 @@ export const PanelLayout = (p: {
   const { form, isEditing } = useForm()
   const token = useCredentials()
 
-  // Get all global glosses / matching tags to display.
-  const { cherokeeRepresentation } = usePreferences()
-  const [{ data }] = Dailp.useGlossaryQuery({
-    variables: { system: cherokeeRepresentation },
-  })
-
-  if (!data) {
-    return <p>Loading...</p>
-  }
-
-  // Get all the tags except for those which are empty/undefined.
-  const allTags = data.allTags.filter((tag) => tag.tag !== "")
-  const groupedTags = groupBy(allTags, (t) => t.morphemeType)
-
-  // Creates a selectable option out of each functional tag, and groups them together by morpheme type.
-  const options: GroupedOption[] = Object.entries(groupedTags).map(
-    ([group, tags]) => {
-      return {
-        label: group,
-        options: tags.map((tag) => {
-          return {
-            // Value is a custom type to track data of a morpheme's gloss in its string form and its matching tag, if there is one.
-            value: tag.tag,
-            label: tag.title,
-          }
-        }),
-      }
-    }
-  )
-
   let panel = null
 
-  // Display the paragraph panel if the segment type is a word (AnnotatedForm).
   if (p.segment.__typename === "AnnotatedForm") {
     panel = (
       <>
@@ -124,11 +83,7 @@ export const PanelLayout = (p: {
         {/* Renders audio recording. */}
         {isEditing ? (
           <Form {...form}>
-            <PanelContent
-              panel={PanelType.EditWordPanel}
-              word={p.segment}
-              options={options}
-            />
+            <PanelContent panel={PanelType.EditWordPanel} word={p.segment} />
           </Form>
         ) : (
           <PanelContent panel={PanelType.WordPanel} word={p.segment} />
@@ -136,7 +91,6 @@ export const PanelLayout = (p: {
       </>
     )
   } else if (p.segment.__typename === "DocumentParagraph") {
-    // Display the paragraph panel if the segment type is a paragraph.
     panel = <ParagraphPanel segment={p.segment} setContent={p.setContent} />
   }
 
@@ -155,20 +109,10 @@ function WordFeature(p: {
   return <div>{p.word[p.feature]}</div>
 }
 
-// A label associated with a list of options.
-type GroupedOption = {
-  label: string
-  options: {
-    value: string
-    label: string
-  }[]
-}
-
 /** Dispatches to the corresponding panel type to render a normal word panel or an editable word panel. */
 export const PanelContent = (p: {
   panel: PanelType
   word: Dailp.FormFieldsFragment
-  options: GroupedOption[]
 }) => {
   // what should be used to render word features? eg, syllabary, commentary, etc.
   const PanelFeatureComponent =
@@ -193,7 +137,7 @@ export const PanelContent = (p: {
         <PanelFeatureComponent
           word={p.word}
           feature={"romanizedSource"}
-          label="Simple Phonetics"
+          label="Romanized Source"
         />
       )}
     </>
@@ -201,7 +145,6 @@ export const PanelContent = (p: {
 
   const translation = (
     <>
-      {/* If the english gloss string is not empty, then display it. */}
       {p.word.englishGloss[0] !== "" && (
         <PanelFeatureComponent
           word={p.word}
@@ -217,24 +160,17 @@ export const PanelContent = (p: {
   // Contains components rendering a word's segments and its english translation.
   const wordPartsContent = (
     <>
+      <VerticalMorphemicSegmentation
+        cherokeeRepresentation={cherokeeRepresentation}
+        segments={p.word.segments}
+      />
+
       {p.panel === PanelType.WordPanel ? (
-        <VerticalMorphemicSegmentation
-          cherokeeRepresentation={cherokeeRepresentation}
-          segments={p.word.segments}
-        />
-      ) : (
-        <EditSegmentation segments={p.word.segments} options={p.options} />
-      )}
-
-      {/* Since editing translations is not yet supported, just display the translation for now. */}
-      <div style={{ display: "flex" }}>‘{p.word.englishGloss}’</div>
-
-      {/* {p.panel === PanelType.WordPanel ? (
         <div style={{ display: "flex" }}>‘{translation}’</div>
       ) : (
         // should this be a call to the parameterized component as well?
         <>{translation}</>
-      )} */}
+      )}
     </>
   )
 
@@ -292,80 +228,6 @@ export const PanelContent = (p: {
         />
       )}
     </>
-  )
-}
-
-// An editable view of a word's parts / segments.
-const EditSegmentation = (p: {
-  segments: Dailp.FormFieldsFragment["segments"]
-  options: GroupedOption[]
-}) => {
-  const { form } = useForm()
-
-  return (
-    <table className={css.tableContainer}>
-      <tbody>
-        {p.segments.map((segment, index) => (
-          <tr style={{ display: "flex" }}>
-            <td className={css.editMorphemeCells}>
-              {/* This is disabled at the moment to be fully implemented later. */}
-              <FormInput
-                {...form}
-                disabled
-                className={formInput}
-                name={["word", "segments", index.toString(), "morpheme"]}
-              />
-            </td>
-            <td className={css.editGlossCells}>
-              {/* Displays global glosses and allows user to create custom glosses on keyboard input. */}
-              <EditGloss
-                // TODO: this key will need to be changed later since a morpheme can be changed
-                key={segment.morpheme}
-                morpheme={segment}
-                index={index}
-                options={p.options}
-              />
-            </td>
-          </tr>
-        ))}
-      </tbody>
-    </table>
-  )
-}
-
-// Component that allows editing of a morpheme's gloss. Users can enter a custom gloss or select from global glosses / functional tags.
-const EditGloss = (props: {
-  morpheme: Dailp.FormFieldsFragment["segments"][0]
-  index: number
-  options: GroupedOption[]
-}) => {
-  const { form } = useForm()
-
-  // Handles gloss selection and creation of new glosses.
-  const handleChange = (
-    newValue: OnChangeValue<{ value: string; label: string }, false>
-  ) => {
-    if (newValue?.value) {
-      const newMorpheme: Dailp.FormFieldsFragment["segments"][0] = {
-        ...props.morpheme,
-        gloss: newValue.value,
-      }
-
-      // Updates current list of morphemes to include one with a matching tag,
-      // or one with a custom gloss.
-      form.update(["word", "segments", props.index], newMorpheme)
-    }
-  }
-
-  return (
-    <CustomCreatable
-      onChange={handleChange}
-      options={props.options}
-      defaultValue={{
-        value: props.morpheme.gloss,
-        label: props.morpheme.matchingTag?.title ?? props.morpheme.gloss,
-      }}
-    />
   )
 }
