+++ website/src/renderer/PageShell.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { HelmetData } from "react-helmet"
 import { Provider as ReakitProvider } from "reakit"
 import { Client as GraphQLClient, Provider as GraphQLProvider } from "urql"
-import type { PageContextBuiltIn } from "vite-plugin-ssr/types"
+import type { PageContextBuiltIn } from "vite-plugin-ssr"
 import type { PageContextBuiltInClient } from "vite-plugin-ssr/client"
 import { UserProvider } from "src/auth"
 import { CollectionProvider } from "src/pages/edited-collections/edited-collection-context"
