+++ website/src/segment.tsx
@@ -7,8 +7,8 @@ import * as Dailp from "src/graphql/dailp"
 import { DocumentContents } from "src/pages/documents/document.page"
 import { std } from "src/style/utils.css"
 import { CleanButton } from "./components"
-import { useForm } from "./edit-word-form-context"
 import { Environment, deploymentEnvironment } from "./env"
+import { useForm } from "./form-context"
 import { PanelDetails } from "./panel-layout"
 import * as css from "./segment.css"
 import { BasicMorphemeSegment, LevelOfDetail } from "./types"
