+++ website/vite.config.ts
@@ -5,10 +5,12 @@ import postcssPresetEnv from "postcss-preset-env"
 import { defineConfig } from "vite"
 import checker from "vite-plugin-checker"
 import ssr from "vite-plugin-ssr/plugin"
+import tsconfigPaths from "vite-tsconfig-paths"
 
 export default defineConfig({
   plugins: [
     react(),
+    tsconfigPaths(),
     vanillaExtractPlugin(),
     // Only check types in development mode.
     process.env.NODE_ENV === "development"
@@ -22,9 +24,6 @@ export default defineConfig({
       },
     }),
   ],
-  ssr: {
-    noExternal: ['lodash']
-  },
   css: {
     postcss: {
       plugins: [
